+++ src/main/java/com/botpulse/demo/repository/SubModuleRepository.java
@@ -1,11 +1,15 @@
 package com.botpulse.demo.repository;
 
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 import com.botpulse.demo.entity.SubModule;
 
 
 @Repository
-public interface SubModuleRepository extends JpaRepository<SubModule, Long> {
+public interface SubModuleRepository extends JpaRepository<SubModule, Integer> {
+
 
 }
